started demo file

+++ simulation/Vector.h
@@ -2,7 +2,8 @@
 
 typedef struct Vector
 {
-    float x;
-    float y;
-    float z;
+    double x;
+    double y;
+    double z;
 } Vector;
+
